src/products/models.py
from django.db import models

from utils.mixins import UUIDMixin


class Category(UUIDMixin):
    name = models.CharField(max_length=50, unique=True)
    parent = models.ForeignKey(
        'self', default=None, null=True, blank=True, on_delete=models.SET_NULL, related_name='sub_categories'
    )

    def __str__(self):
        return self.name

    @property
    def is_sub_category(self):
        if self.parent:
            return True
        return False

    @property
    def is_parent_category(self):
        if self.sub_categories.first():
            return True
        return False
